% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommenters.R
\name{getCommenters}
\alias{getCommenters}
\title{Get usernames of top-level Reddit commenters}
\usage{
getCommenters(thread)
}
\arguments{
\item{thread}{The unique portion of a reddit url beginning with /r/ for which you would like to identify the top-level commenters}
}
\value{
A vector of Reddit usernames belonging to the top-level commenters for the given thread
}
\description{
Get usernames of top-level Reddit commenters
}
