% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComments.R
\name{getComments}
\alias{getComments}
\title{Get top-level comments for a given Reddit thread}
\usage{
getComments(thread)
}
\arguments{
\item{thread}{The unique portion of a reddit url beginning with /r/ for which you would like to identify the top-level commenters}
}
\value{
A data frame containing the author, subreddit, thread, comment id, time created in UTC, the comment text, the number of direct replies, the up votes received, the down votes received, the post score, status of whether or not the comment can be gilded, and whether or not it is already gilded
}
\description{
Get top-level comments for a given Reddit thread
}
