% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentThreads.R
\name{getRecentThreads}
\alias{getRecentThreads}
\title{Get Reddit Top Threads}
\usage{
getRecentThreads(subreddit)
}
\arguments{
\item{subreddit}{The name of a subreddit as it appears in its url}
}
\value{
A vector of urls beginning with /r/ for the top posts from the last 24 hours for the given subreddit
}
\description{
Get Reddit Top Threads
}
