% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentThreadsNew.R
\name{getRecentThreadsNew}
\alias{getRecentThreadsNew}
\title{Get New Reddit Threads}
\usage{
getRecentThreadsNew(subreddit)
}
\arguments{
\item{subreddit}{The name of a subreddit as it appears in its url}
}
\value{
A vector of urls beginning with /r/ for the latest posts for the given subreddit
}
\description{
Get New Reddit Threads
}
