% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserInfo.R
\name{getUserInfo}
\alias{getUserInfo}
\title{Gather Reddit user info}
\usage{
getUserInfo(user)
}
\arguments{
\item{user}{A Reddit username given as a string}
}
\value{
A data frame containing the user's username, account creation date in UTC, link karma, comment karma, Reddit Gold status, moderator status, verified status, verified email status, number of comments (up to 100), and number of submissions (up to 100)
}
\description{
Gather Reddit user info
}
