% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giveGold.R
\name{giveGold}
\alias{giveGold}
\title{Give gold to posts, comments, and accounts}
\usage{
giveGold(id)
}
\arguments{
\item{id}{The fullname of the post (t3_), comment (t1_), or account (t2_) being rewarded}
}
\value{
The status message from the POST function
}
\description{
Give gold to posts, comments, and accounts
}
