% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeToTreatment.R
\name{randomizeToTreatment}
\alias{randomizeToTreatment}
\title{Randomly assign subjects to treatment conditions}
\usage{
randomizeToTreatment(n, treatments = c("C", "T"), weights)
}
\arguments{
\item{n}{The number of subjects}

\item{weights}{An optional argument that allows users to change the ratio of subjects in each condition}

\item{treatment}{A vector of available treatment options. Defaults to c("C", "T")}
}
\value{
A vector of treatment assignments
}
\description{
Randomly assign subjects to treatment conditions
}
\examples{
randomizeToTreatment(20, c("C", "T1", "T2"))
randomizeToTreatment(20, c("C", "T1", "T2"), c(0.5, 0.25, 0.25))
}
