% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upvoteContent.R
\name{upvoteContent}
\alias{upvoteContent}
\title{Upvote posts and comments}
\usage{
upvoteContent(id, dir)
}
\arguments{
\item{id}{The fullname of the post (t3_) or comment (t1_) being upvoted}

\item{dir}{The vote direction; one of (1, 0, -1)}
}
\value{
The status message from the POST function
}
\description{
Upvote posts and comments
}
